Update and rename Primes'.hs to Primes.hs
This is an older version before I understand how a library was supposed to necessarily be structured. So there is a main function that prints out what all the library can do.

Next update will be the updated version, which rewrites a lot of the inner logic of functions, and adds a Haddock style comment ahead of function names. Main will be deleted, and documentation moved to comments.