Update prjEuler36.hs
Realized I didn't need to break it into cases, and can just generate two palindromes at a time for every number [1..999].

1 -> [1, 11]
12 -> [121, 1221].

The older variant (generating palindrome by length and then concatenating) is nicer in that it is already sorted, but since we are summing at the end anyways, order does not matter. And the code is much shorter.